pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.4"
description = "Gendiff is a program that determines the difference between two data structures."
authors = [
  "Dmitry Samoylenko <dimsamoylenko@gmail.com>"
]
packages = [
  { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
wemake-python-styleguide = "^0.15.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
